add utils to plot to break circular dependency
